pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trade_advisor"
version = "0.1.0"
description = "Local-first quant trade advisor with backtesting, walk-forward, and ML"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "You" }]

dependencies = [
    # Data
    "yfinance>=0.2.40",
    "pandas>=2.2",
    "numpy>=1.26",
    "pyarrow>=15.0",
    "duckdb>=1.0",
    # Indicators
    "pandas-ta>=0.3.14b",
    # Backtesting
    "vectorbt>=0.26",
    # Tracking
    "mlflow>=2.13",
    # UI + CLI
    "streamlit>=1.35",
    "plotly>=5.22",
    "typer>=0.12",
    "rich>=13.7",
    # Config
    "pydantic>=2.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-playwright>=0.5",
    "ruff>=0.5",
    "black>=24.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "types-pyyaml",
    "pandas-stubs",
]

[project.scripts]
ta = "trade_advisor.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/trade_advisor"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "integration: hits the network (yfinance); skipped in CI by default",
    "e2e: requires running Streamlit server; use --browser chromium",
    "slow: tests that take >1s",
]
